% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tensors.R
\name{calc_Qinvariant}
\alias{calc_Qinvariant}
\title{Q-invariant}
\usage{
calc_Qinvariant(
  u_fld,
  v_fld,
  w_fld,
  lev_p,
  lat = NULL,
  dx = 0.25,
  dy = 0.25,
  mode = "lonlat"
)
}
\arguments{
\item{u_fld}{zonal velocity field [m/s]}

\item{v_fld}{meridional velocity field [m/s]}

\item{w_fld}{vertical velocity field [m/s]}

\item{lev_p}{vector containing pressure levels [Pa]}

\item{lat}{vector containing latitude}

\item{dx}{x resolution in the corresponding unit (e.g. 0.25 degree for ERA5 with \code{mode='lonlat'} or e.g. 1000 m in cartesian coordinates with \code{mode='cartesian'})}

\item{dy}{y resolution in the corresponding unit (e.g. 0.25 degree for ERA5 with \code{mode='lonlat'} or e.g. 1000 m in cartesian coordinates with \code{mode='cartesian'})}

\item{mode}{use 'lonlat' if the data is given on a lon-lat-grid or 'cartesian' if the data is given on an equidistant cartesian grid}
}
\value{
Q-invariant [1/s^2]
}
\description{
Calculates the Q-invariant := 1/2 * ( frobenius_norm(Omega)^2 - frobenius_norm(S)^2 )
}
\examples{
myfile=system.file("extdata", "era5_storm-zeynep.nc", package = "meteoEVT")
data = readin_era5(myfile)
S=calc_Qinvariant(data$u,data$v,data$w,data$lev,lat=data$lat)

}
