% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_operators.R
\name{crossprod}
\alias{crossprod}
\title{cross product}
\usage{
crossprod(fld1, fld2)
}
\arguments{
\item{fld1}{field 1 with dimensions (lon,lat,p,3)}

\item{fld2}{field 2 with dimensions (lon,lat,p,3)}
}
\value{
field containing the cross product
}
\description{
Calculates the cross product of two given 3d vector fields
}
