% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_operators.R
\name{df_dz}
\alias{df_dz}
\title{df_dz}
\usage{
df_dz(fld, rho, plev = 5000)
}
\arguments{
\item{fld}{field with dimensions (lon,lat,p)}

\item{rho}{field with dimensions (lon,lat,p) for density or a scalar rho (for constant density)}

\item{plev}{a scalar containing the p resolution (if equidistant) or a vector containing pressure levels in Pa (for non-equidistant)}
}
\value{
field containing the partial derivative w.r.t. z
}
\description{
Calculates the z derivative
}
