% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{fill_horiz}
\alias{fill_horiz}
\title{Plotting a xy domain with custom boundaries, colour paletts and optional world map}
\usage{
fill_horiz(
  x,
  y,
  fld,
  levels = 1:100,
  main = "",
  worldmap = TRUE,
  legend_loc = "topright",
  legend_title = "",
  legend_only = FALSE,
  Lab = NULL,
  ...
)
}
\arguments{
\item{x}{array containing x-axis values (e.g. longitude)}

\item{y}{array containing y-axis values (e.g. latitude)}

\item{fld}{field (which should be plotted) with dimensions (x,y)}

\item{levels}{levels for colour bar}

\item{main}{character containing main title of the plot}

\item{worldmap}{should the world map contours be plotted (default TRUE)}

\item{legend_loc}{location of legend}

\item{legend_title}{character containing legend title}

\item{legend_only}{logical TRUE only legend should be pltted, or FALSE everything should be plotted (default)}

\item{Lab}{lab palette from type colorRampPalette}

\item{...}{additional graphic parameters}
}
\value{
no return
}
\description{
Plotting a xy domain with custom boundaries, colour paletts and optional world map
}
