% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_operators.R
\name{frobenius_norm}
\alias{frobenius_norm}
\title{Frobenius norm}
\usage{
frobenius_norm(mat)
}
\arguments{
\item{mat}{}
}
\value{
Frobenius norm
}
\description{
Calculates the Frobenius norm of a given matrix or array
}
