% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_evt.R
\docType{package}
\name{meteoEVT-package}
\alias{meteoEVT-package}
\title{Introduction}
\description{
Energy-Vorticity Theory (EVT) is the fundamental theory to describe processes in the Earth's atmosphere by combining conserved quantities from hydrodynamics and thermodynamics. This package provides functions to calculate many energetic and vortical quantities, like potential vorticity, Bernoulli function and dynamic state index, for given gridded data, like ERA5 reanalysis. These quantities can be used, e.g., to identify atmospheric fronts, for which are separate functions provided here.
}
\details{
Phenomenons in the Earth's atmosphere, like tropical hurricanes or extratropical cyclones, can adequately be chararcterized by a combination of energetic and vortical quantities. These quantities can also be used for a consistent theoretical description of these phenomenons. This package provides functions to calculate Bernoulli function, vorticity, enstrophy, helicity, Lamb vector and potential vorticity based on given gridded data sets. Addiotionally, by using energy-vortex theory an adiabatic, stationary and invisicid basic state of the Earth's atmosphere can be derived, which is itself a solution of the primitive equations. The derivation from this basic state is given by the dynamic state index (DSI), which can be used for the study of, e.g., cyclones and fronts. Recently, the DSI was used to identify atmospheric fronts objectively from reanalysis data and thereby provides an alternative way for front detection. For this purpose, this package provides funtions to calculate the DSI and use it to identify atmospheric fronts. This method can be compared with state-of-the-art front identification methods based on the thermic front parameter or the F diagnostic.
}
\references{
\itemize{
\item Weber and Nevir
\item Parfitt
\item Hewson
\item arxiv
}
}
