% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readin.R
\name{readin_dim}
\alias{readin_dim}
\title{read in dimensions}
\usage{
readin_dim(filename)
}
\arguments{
\item{filename}{name of file to read in}
}
\value{
no return
}
\description{
: reads dimensions of ERA5 data
}
\examples{
myfile=system.file("extdata", "era5_storm-zeynep.nc", package = "meteoEVT")
data_dims = readin_dim(myfile)
}
