% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readin.R
\name{readin_era5}
\alias{readin_era5}
\title{read in ERA5 data}
\usage{
readin_era5(filename)
}
\arguments{
\item{filename}{name of file to read in}
}
\value{
no return
}
\description{
: reads ERA5 data
}
\examples{
myfile=system.file("extdata", "era5_storm-zeynep.nc", package = "meteoEVT")
data = readin_era5(myfile)
}
