% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_operators.R
\name{scalarprod}
\alias{scalarprod}
\title{scalar product}
\usage{
scalarprod(fld1, fld2)
}
\arguments{
\item{fld1}{field 1 with dimensions (lon,lat,p,d)}

\item{fld2}{field 2 with dimensions (lon,lat,p,d)}
}
\value{
field of the scalar product with dimensions (lon,lat,p)
}
\description{
Calculates the scalar product of two given fields
}
